git: add `gcp` alias, and refactor `gcptm`
i wanted echoing to happen automatically for gcptm, and setting PS4 and
`-x` basically does that, but it doesn't kill execution when a command
fails.  setting `-e` actually causes the whole shell to terminate when a
command fails. maybe i could run things in a subshell or something...
anyway, tacking on a `|| return $?` to every line gets the job done