zsh: rolling with some fancy new plugins
i don't know yet if i want to keep them... but i played with fish for a
while, and i decided not to stick with it because it isn't POSIX
compatible, so e.g. it does for loops differently and env var expansions
(like ${SOME_VAR:-SOME_VAR} kinds of things) differently, and i didn't
really want to get in the habit of learning non-portable ways of doing
things.  but fish had this fancy syntax highlighting and autosuggestion
thing going on, so i added very similar plugins to zsh to try them out
for a while.